+++ app/Http/Controllers/Api/KaryaUPNController.php
@@ -39,13 +39,13 @@ class KaryaUPNController extends Controller
                 if ($result->isEmpty()) {
                     return response()->json([
                         'success' => false,
-                        'message' => 'Data KaryaUPN Kosong!',
+                        'message' => 'Data Karya UPN Kosong!',
                         'code'    => 404,
                     ]);
                 }
                 return response()->json([
                     'success' => true,
-                    'message' => 'OK',
+                    'message' => 'Berhasil',
                     'code'    => 200,
                     'data'    => $result
                 ]);
@@ -105,7 +105,7 @@ class KaryaUPNController extends Controller
 
                         return response()->json([
                             'success' => true,
-                            'message' => 'OK',
+                            'message' => 'Berhasil',
                             'code'    => 200,
                             'data'    => $result
                         ]);
@@ -144,13 +144,13 @@ class KaryaUPNController extends Controller
                 if (!$result) {
                     return response()->json([
                         'success' => false,
-                        'message' => 'Data KaryaUPN Kosong!',
+                        'message' => 'Data Karya UPN Kosong!',
                         'code'    => 404,
                     ]);
                 }
                 return response()->json([
                     'success' => true,
-                    'message' => 'OK',
+                    'message' => 'Berhasil',
                     'code'    => 200,
                     'data'    => $result
                 ]);
@@ -193,31 +193,30 @@ class KaryaUPNController extends Controller
                 ]);
             }
             else {
-
+                $result = $this->service->getOne($id);
                 if (!$result) {
                     return response()->json([
                         'success' => false,
-                        'message' => 'Service Error',
-                        'code'    => 500,
+                        'message' => 'Data Karya UPN Tidak Ditemukan!',
+                        'code'    => 404,
                     ]);
                 } else {
                     $result = $this->service->update($this->request, $id);
-                    // $result = $this->service->store($this->request);
-                        if (!$result) {
-                            return response()->json([
-                                'success' => false,
-                                'message' => 'Service Error',
-                                'code'    => 500,
-                            ]);
-                        }
-
+                    if (!$result) {
                         return response()->json([
-                            'success' => true,
-                            'message' => 'OK',
-                            'code'    => 200,
-                            'data'    => $result
+                            'success' => false,
+                            'message' => 'Service Error',
+                            'code'    => 500,
                         ]);
                     }
+
+                    return response()->json([
+                        'success' => true,
+                        'message' => 'Berhasil',
+                        'code'    => 200,
+                        'data'    => $result
+                    ]);
+                }
             }
         } catch (\Throwable $th) {
             dd('Controller error ' . $th->getMessage());
@@ -247,11 +246,20 @@ class KaryaUPNController extends Controller
                 ]);
             }
             else {
+                $result = $this->service->getOne($id);
+                if (!$result) {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'Data Karya UPN Tidak Ditemukan!',
+                        'code'    => 404,
+                    ]);
+                }
+
                 $result = $this->service->delete($id);
 
                 return response()->json([
                     'success' => true,
-                    'message' => 'OK',
+                    'message' => 'Berhasil',
                     'code'    => 200,
                 ]);
             }
